plugins/aws-agentcore-langgraph/examples/baml-extraction/pyproject.toml
[project]
name = "baml-extraction-demo"
version = "0.1.0"
description = "BAML extraction demo for AWS Bedrock AgentCore memory events"
requires-python = ">=3.10"
dependencies = [
    "baml-py>=0.85.0",
    "python-dotenv>=1.0.0",
    "typing_extensions>=4.0.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]
